refactor: move types to networkConfig
